added check if anime exists

+++ main.py
@@ -13,88 +13,107 @@ while True:
     print("5. print all animes in table")
 
     choice = input("Pick 1-5 to do something: ").strip()
+    print("\n")
   
 
     match choice:
             case "1": #add a new anime
                 anime_name = input("Enter the anime name: ").strip()
 
+                if anime.anime_real(anime_name) == False: #check if the anime exists
+                    print("Anime not found in database.")
+                    continue
+
                 try:    
                     total_episodes = int(input("Enter the total number of episodes: "))
-                    if (total_episodes < 1 or total_episodes > 3000):
-                        print("Total episodes must be between 1 and 3000.")
-                        continue
+                    watched_episodes = int(input("Enter the number of episodes you have watched: "))
                 except ValueError:
                     print("Total episodes must be numbers.")
                     continue
-
-
-                try:
-                    watched_episodes = int(input("Enter the number of episodes you have watched: "))
-                    if (watched_episodes < 0 or watched_episodes > 3000):
-                        print("Total episodes entered must be between 0 and 3000")
-                        continue
-                except ValueError:
-                    print("Total episodes in series must be number values.")
-                    continue
                 
                 if total_episodes < 0 or watched_episodes < 0:
-                    print("Sorry! The total number of episodes can't be less than 0.")
+                    print("Total episodes can't be less than 0.")
                     continue
 
                 if watched_episodes > total_episodes:
-                    print("Sorry! The watched number of episodes can't be less than 0.")
+                    print("Watched episodes can't be more than total episodes.")
                     continue
 
                 this_anime = anime.add_anime(anime_name, total_episodes, watched_episodes) 
 
+                print("Anime shown successfully.")
                 print(this_anime)
                 print("\n")
+
+#-----------------------------------------------------------------------
 
             case "2":  #update an anime 
                anime_name = input("Enter the anime name: ").strip()
                
+               if anime.anime_real(anime_name) == False: #check if the anime exists
+                    print("Anime not found in database.")
+                    continue
+               
                try:    
                     watched_episodes = int(input("Enter the number of episodes you have watched: "))
-                    if (watched_episodes < 0 or watched_episodes > 3000):
-                        print("Total episodes must be between 1 and 3000")
-                        continue
                except ValueError:
-                    print("Total episodes in series must be number values.")
+                    print("Total episodes must be numbers.")
                     continue
                
                if watched_episodes < 0:
-                    print("Sorry! The watched number of episodes can't be less than 0.")
+                    print("Watched episodes can't be less than 0.")
                     continue
                
                updated_anime = anime.update_anime(anime_name,watched_episodes)
-               print(updated_anime)
+               print("Anime has updated sucessfully.")
                print("\n")
                
+#-----------------------------------------------------------------------
+
             case "3": #delete an anime
                 anime_name = input("Enter the anime name: ").strip()
+
+                if anime.anime_real(anime_name) == False: #check if the anime exists
+                    print("Anime not found in database.")
+                    continue
+
+
                 deleteanime = anime.delete_anime(anime_name)
-                print(deleteanime)
+                print("Anime deleted successfully.")
                 print("\n")
 
+
+#-----------------------------------------------------------------------
             case "4": #whats the progress on all anime?
                 anime_name = input("Enter the anime name: ").strip()
+
+                if anime.anime_real(anime_name) == False:  #check if the anime exists
+                    print("Anime not found in database.")
+                    continue
+
                 progressanime = anime.anime_progress(anime_name)
                 completedanime = anime.anime_completed(anime_name)
                 print (progressanime)
                 print (completedanime)
                 print("\n")
-                       
+
+#-----------------------------------------------------------------------
+
             case "5":  #print anime name list
                 anime.anime_titles()
                 print("\n")
 
+
+#-----------------------------------------------------------------------
+
             case default:
-                print("Use numbers punk ass bitch")
+                print("use numbers punk ass bitch")
+                print("\n")
 
 
     continue_choice = input("Do you want to continue adding anime? (Y/N): ").strip().lower()
     if continue_choice != ("y"):
         print("No more anime will be added.  Thank you for your time.")
+        print("\n")
         break
